[mod] coding refactoring
