+++ CodeBaseProject/Domain/Infrastructures/LocalService.swift
@@ -1,0 +1,12 @@
+//
+//  LocalService.swift
+//  CodeBaseProject
+//
+//  Created by HaiKaito on 03/04/2022.
+//
+
+import Foundation
+
+// Use for local data
+//protocol DataStoreServiceProtocol: UseCasesProtocol {
+//}
