+++ CodeBaseProject/Domain/Infrastructures/UseCaseConstructor.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-class UseCaseConstructor: UseCasesProtocol {
+class UseCaseConstructor {
 	private var useCaseProvider: UseCaseProviderProtocol
 	
 	// MARK: - ================================= Init =================================
@@ -35,40 +35,5 @@ extension UseCaseConstructor {
 	// Update
 	func update(inputUrl: UrlInputUserInfo, inputBody: BodyInputUserInfo) async -> FinalResult<Bool> {
 		return await useCaseProvider.user.update(inputUrl: inputUrl, inputBody: inputBody)
-	}
-}
-
-class NetworkService: UseCaseConstructor {
-}
-
-private class DemoPreview {
-	func execute() async throws {
-		let networkProvider = NetProvider()
-		let useCaseProvider = NetUseCaseProvider(networkProvider)
-		let networkService = NetworkService(useCaseProvider: useCaseProvider)
-		
-		//		do {
-		//
-		//			defer {
-		//				print("do complete..")
-		//			}
-		//
-		//			let result = try await networkService.userInfo(inputUrl: UrlInputUserInfo(userId: "123", gender: nil))
-		//			print(result)
-		//
-		//		} catch is ErrorPointer { //generic error
-		//			print("error")
-		//		} catch { // else
-		//			print("error")
-		//		}
-		
-		let result = await networkService.userInfo(inputUrl: UrlInputUserInfo(userId: "123", gender: nil))
-		switch result {
-		case .success(let value):
-			print(value)
-			break
-		case .failure(let error):
-			print(error)
-		}
 	}
 }
